requirements.txt
# AI标书系统核心依赖
Flask==2.3.3
Flask-CORS==4.0.0
Flask-WTF==1.2.1           # CSRF保护
Flask-Compress==1.14       # Gzip/Brotli压缩
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0           # 生产环境WSGI服务器

# 数据库相关
# sqlite3 是 Python 标准库，无需安装

# 文档解析相关
PyMuPDF>=1.23.0           # PDF解析
Pillow>=10.0.0            # 图片处理 (PyMuPDF转换PDF时需要)
python-docx>=0.8.11       # Word文档解析
mammoth>=1.0.0            # Word转HTML (保留格式)
pdfplumber>=0.9.0         # PDF表格提取
openpyxl>=3.1.0           # Excel文件处理

# 文本处理和分块
langchain>=0.1.0          # 文本分块工具
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0           # token计数
jieba>=0.42.1             # 中文分词

# 向量化和搜索
faiss-cpu>=1.7.4          # 向量检索引擎
# sentence-transformers>=2.2.2  # 句子转换器（已改用OpenAI API，无需本地模型）
# transformers>=4.30.0      # Transformer模型（已改用OpenAI API，无需本地模型）
# torch>=2.0.0              # PyTorch（已改用OpenAI API，无需本地模型，1-2GB很慢）
numpy>=1.24.0             # 数值计算
scikit-learn>=1.3.0       # 机器学习工具

# 异步处理（可选，当前未使用）
# celery>=5.3.0             # 异步任务队列（暂时注释，部署时可能不兼容）
# redis>=4.5.0              # 缓存和消息队列（暂时注释）

# 文本清洗和处理
beautifulsoup4>=4.12.0    # HTML解析
lxml>=4.9.0               # XML解析器
regex>=2023.6.3           # 高级正则表达式

# 日志和监控
loguru>=0.7.0             # 增强日志系统

# 开发工具
pytest>=7.4.0            # 单元测试
black>=23.0.0             # 代码格式化

# 认证和安全
PyJWT>=2.8.0              # JWT token生成和验证
bcrypt>=4.1.0             # 密码加密